fix error in matcher-siren when deploying with monitoring disabled
