+++ webofneeds/won-matcher-siren/src/main/java/won/matcher/siren/actor/SirenMonitoringMatcherActor.java
@@ -23,7 +23,7 @@ import java.io.IOException;
  */
 @Component
 @Scope("prototype")
-public class SirenMonitoringMatcherActor extends SirenMatcherActor
+public class SirenMonitoringMatcherActor extends AbstractSirenMatcherActor
 {
 
   @Override
